safari-600.7-branch/LayoutTests/media/video-click-dblckick-standalone-expected.txt


This tests that clicking on a standalone video will pause and double-clicking will play.

PASS: video is paused.
PASS: video is playing
